Warmup Flare cache after 1.5 seconds on the welcome screen.
