+++ lib/main.dart
@@ -17,26 +17,6 @@ void main() {
   // Don't prune the Flare cache, keep loaded Flare files warm and ready
   // to be re-displayed.
   FlareCache.doesPrune = false;
-  // Populate Flare cache to prevent janks.
-  var warmup = [
-    "assets/flare/Coin.flr",
-    "assets/flare/Designer.flr",
-    "assets/flare/Joy.flr",
-    "assets/flare/ProgramManager.flr",
-    "assets/flare/Sourcerer.flr",
-    "assets/flare/Tester.flr",
-    "assets/flare/TheArchitect.flr",
-    "assets/flare/TheHacker.flr",
-    "assets/flare/TheJack.flr",
-    "assets/flare/TheRefactorer.flr",
-    "assets/flare/Users.flr",
-    "assets/flare/UXResearcher.flr",
-  ];
-  for (final flare in warmup) {
-    cachedActor(rootBundle, flare).then((actor) {
-      print("Loaded $flare...");
-    });
-  }
 
   runApp(MyApp());
 }
