+++ lib/src/welcome_screen.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:dev_rpg/src/game_screen/character_style.dart';
 import 'package:dev_rpg/src/shared_state/game/world.dart';
 import 'package:dev_rpg/src/style.dart';
+import 'package:dev_rpg/src/widgets/flare/warmup_flare.dart';
 import 'package:dev_rpg/src/style_sphinx/sphinx_screen.dart';
 import 'package:dev_rpg/src/widgets/buttons/welcome_button.dart';
 import 'package:dev_rpg/src/widgets/flare/start_screen_hero.dart';
@@ -17,6 +18,8 @@ class WelcomeScreen extends StatefulWidget {
 class _WelcomeScreenState extends State<WelcomeScreen> {
   CharacterStyle hero;
   Timer _swapHeroTimer;
+  final Timer _warmupTimer =
+      Timer(const Duration(milliseconds: 1500), warmupFlare);
   @override
   void initState() {
     chooseHero();
@@ -35,6 +38,7 @@ class _WelcomeScreenState extends State<WelcomeScreen> {
   void dispose() {
     super.dispose();
     _swapHeroTimer?.cancel();
+    _warmupTimer.cancel();
   }
 
   static const double _horizontalPadding = 33.0;
